server/requirements.txt
flask
numpy
opencv-python
pillow
pymongo
dnspython
